Create Hackerearth "to buy or not buy"
Trick - To check whether an edge is included in a minimum spanning tree or not (if there are many edges of same weight), take two nodes of the corresponding edge, build the graph excluding the edge and taking only those edges whose weight is below the selected edge weight, then select a node out of both of them and finally run a dfs on the selected node. Now if the other node is included in the dfs means that there were other paths too to the reach the other node. And finally its clear that the included edge was definitely not going to be included in the MSP. So is the ans.